chart(responsive): guard for when no handler is attached
The only way I found to check this is to access the method that will be called and if it throws - then it is null and you can't call it. I prefer that over a try-catch around the invoke because the handlers should implement their own error handling in my view.