Brought in changes to SqlBuilderExtensions (non-breaking), and provided an example of the additional Where clause being useful.
The main sales page example shows pulling all sales as normal, with the products. Then each sales page specifying a product will automatically filter to sales relating to that product ID, without applying any filter to the Grid. It emulates showing more info for a product, so that you only want to see information on that product.